Improving display format for strings and FourCCs
String and FourCC field values are now wrapped in quotation marks,
which makes it easier to see where they begin and end. For
example, when wrapping a "fmt " chunk ID in quotes, it makes it
possible to see that there is a space character at the end that
otherwise would not be visible. (Note: before this commit FourCC
values were already wrapped in quotation marks when shown in a
chunk title, but not when shown in the table listing the chunk's
fields).

In addition, character codes are now shown in place of non-display
characters, which also makes it easier to see the actual contents
of a string/FourCC.

As part of making this change, the `gsub("\n\n", "")` that was
applied to display values is being removed. I'm not sure why that
was originally added, but I'm guessing it was in response to a
file that contained many newline characters in a string value.
With the new display format, that change shouldn't be needed,
because newline characters will be shown as "\n" instead of
actually adding a newline character to the output. And besides
that, it seems like a good idea to avoid one-off formatting logic
like this in general.