Debug: Removing column alignment of arguments
Although this alignment made some of the code a little easier to
read, it was inconsistently applied, and this type of thing can
make diffs noisier. (I.e. if a field is added/modified, resulting
in the length of the longest field name changing, other lines have
to also change to keep the alignment in tact).