Debug: Better mechanism for reading fields
This replaces the `read_bytes` method with a new `FieldReader`
class. This class provides named methods for reading fields, such
as `read_guid`, `read_fourcc`, or `read_uint32`, which removes the
need for the caller to manually deal with an `unpack` string. This
allows the caller to focus on the semantic type of the field,
rather than the implementation detail of how to extract the value
from a set of bytes. For example, `field_reader.read_guid` seems
more clear than `read_bytes("H32")`.

In addition, the new class has a few other benefits:

* Reduces the need to directly use the global `FILE` variable,
  which is a step towards allowing this variable to be defined in
  a more narrow scope.
* Reduces repetition of common logic.
* Lays the groundwork for improved functionality in the future.