Debug: Fixing logic to determine cue point count
The "cue point count" field is now used to determine how many cue
points to read, instead of a (potentially incorrect) calculation
based on the number of bytes in the chunk. As a side effect, if
there are extra bytes at the end of the chunk, they will now be
properly displayed as extra bytes, instead of being interpreted as
the first bytes of the next chunk.