Debug: Fixing "LIST" child chunk reading logic
The previous check for there being more than 1 byte remaining
seems wrong; it seems like the search for child chunks should
continue as long as there are any bytes remaining. (In practice,
there would always need to be 8 or more bytes left for there to be
a valid child chunk left, but this script intentionally doesn't
attempt to do that type of validation).

As a side effect, it seems like the check for extra bytes at the
end of the chunk will no longer ever succeed, so that is being
removed as well. Conceptually, it doesn't seem possible for a
"LIST" chunk to have extra bytes at the end, so that seems like an
improvement.