Debug: Parsing fields in `ltxt` chunks
Although the data in an `ltxt` chunk was previously displayed as a
single string, it actually contains several structured fields
followed by optional text. The structured fields are now parsed
and displayed.

The optional text is just shown as raw bytes to avoid a
potentially wrong implementation. The problem is that the "Code
Page" field seemingly implies that different text encodings are
possible, but the spec doesn't provide more info, and I don't have
a real world example file containing optional text to use for
comparison.