Debug: Properly reading null-terminated strings
The relevant null-terminated string fields will no longer display
the terminating null character (i.e. `\x00`), or any subsequent
characters.

These fields are also now labeled as having the type "C String".
Although the spec uses the term "ZSTR" to refer to this kind of
string, "C String" seems like it might be a more common term...?

Also, the byte count is no longer shown as part of the type (e.g.
"alpha_28" prior to this commit) because it seems both redundant
due to the chunk size field, and also because it was inaccurate
due to the count including the null character (and potentially
characters after the null).