Debug: Closing file if any type of error occurs
Previously, the file would only be closed upon `EOFError` being
raised, but it seems like the file should be closed if _any_ type
of error occurs. (In practice, I believe the file should be
automatically closed anyway when the script exits, so this is more
a matter of hygiene rather than actually preventing an open file
leak).