Debug: Showing skipped byte counts in output
This makes the presence of content that is skipped more explicit.
While this adds clarity when displaying valid files, it seems
especially helpful when a file ends prematurely. Previously, there
wouldn't be any indication of this in the output, but now there
will be via a smaller-than-expected skipped byte count being
shown.