Debug: Standardizing field variable name suffixes
All names for variables that hold the result of a field read out
of the file now have a consistent "_field" suffix, instead of
"_data" or "_bytes" or no suffix at all. This hopefully makes it
more clear which variables represent a full field, rather than a
component of a field (such as the parsed value).