Debug: Show chunk ID even if chunk size is missing
Previously, if a chunk header's ID field was present but the size
field was missing or incomplete, the chunk ID would not be shown
in the output. This gave a misleading impression of the file's
actual contents. After this commit, if a chunk header's ID field
is present, it should always be shown regardless of whether or not
the size field is present.