Correcting "RIFF" chunk sizes in fixtures
The "RIFF" chunk size in each of these fixtures did not match the
actual number of bytes in the chunk. This caused each of them to
have an extra source of error beyond what they were intended to
include. Correcting them should hopefully help prevent future bugs
that could be masked by "did this cause an error" tests passing
for an unintended reason.