Better formatting for bug fix summary descriptions
d74615a7ca4bd77efb854a0cd8c8898ef1c28bf7 changed the formatting of
the bug fix descriptions in the README release notes. The first
sentence of each description was bolded, to make it stand out more
when viewed on GitHub. However, after viewing the result of that,
the bolding doesn't stand out as much as I would like and it seems
like maybe a better solution would be to un-bold each summary
sentence but place it on its own line. We will see!

The CHANGELOG is being reverted to use the original bolding style
from before the aforementioned commit. Since this file is intended
to be viewed as plain text, I'm not sure the formatting has to
match the README exactly. Adding more line breaks might/might not
make it easier to read, but I'm keeping the existing style for
consistency with the other previous release notes in the file.