Fixing bug description in release notes
The previous description was inaccurate because it claimed that
any time a "fmt " chunk had extra bytes at the end, the chunk
extension would have the expected number of bytes or more (for
chunk extension formats that are understood by the gem). That was
incorrect.

As a counterexample, suppose a file in WAVE_FILE_EXTENSIBLE format
has a chunk extension shorter than 22 bytes, followed by extra
bytes, with the parent "fmt " chunk size correctly accounting for
the shortened extension size and extra bytes. In v1.1.1, the error
message when trying to open such a file would (by accident)
correctly state "Not a supported wave file. The format chunk
extension is shorter than expected."