Removing extra parentheses in CHANGELOG
This makes the text in the CHANGELOG file consistent with the text
in README.markdown.

In c67fee60cbeb896719935f7580b1e39fd99e7f98 the parentheses around
this sentence were removed from README.markdown, because they no
longer seemed to be needed due to the updated formatting. However,
they were kept in CHANGELOG because the formatting was not updated
in that file. However, I no longer think the parentheses need to
be present in CHANGELOG either, and the removal also has the
benefit of making the two files more consistent.