Debug: Fixing script to work in Ruby 2.0 - 2.4
The previous implementation relied on the more succinct syntax
added in Ruby 2.5 that allows a block to have a `rescue` clause
without also having a `begin`/`end` wrapper.

Although the debug script is not packaged with the gem, it seems
like a good idea for it to work in the same Ruby versions that the
gem supports, unless there is a good reason for it not to.