Fixing inaccurate code comment
This fixture changed in daaefe760c5851d8c687e617f0941aa07e5bda52
to have a format code of `6` instead of `2`, but this comment was
not updated. (This comment was also left out of
363ecea49273b57ef8b496c172c493c95e5b451b, which made a similar
correction to a different file).

To prevent it from becoming incorrect again in the future, the
comment is being changed to not mention the specific format code
the fixture is supposed to have.