Fixing bug description in release notes
The previous version of this paragraph was misleading because it
incorrectly implied that this special case could apply to files
that had a format code that wasn't `1`.

If the format code is not `1`, then bytes 16 and 17 (0-based)
contain the "extension size" field. For this value to match the
remaining number of bytes in the chunk, the extension has to take
up the entire remainder of the chunk, leaving no room for extra
trailing bytes. Therefore, the special case mentioned in this
paragraph was only possible if the format code was `1`.

The removed "Except if the file was in WAVE_FORMAT_EXTENSIBLE..."
sentence, added in d7bdfb8d2688725ce9deb00fd1b7a6e2910fc396, seems
to have been the result of confusing extra bytes at the end of an
extension (which isn't covered by this bug fix) with extra bytes
at the end of the chunk.