Reordering paragraphs in release notes for clarity
It seems helpful to move the definition of "extra bytes" to the
start of the bug fix description, since otherwise the reader might
be wondering what that means exactly. This also seems important
because other bug fixes in this release also deal with scenarios
that could be interpreted as involving "extra bytes", and could be
confused with this one.