Modifying example to avoid possible confusion
The size of the "fmt " chunk in the example is intentionally being
changed to a value other than 40, to avoid inadvertently implying
that this bug fix only applies to files in WAVE_FORMAT_EXTENSIBLE
format. Together with the final sentence of the previous bug fix,
it seems like the old example size could have led to that
conclusion.