Removing inaccurate statement about field name
The statement that the spec calls this field the "byte rate" seems
wrong. The spec calls this field "dwAvgBytesPerSec", and I can't
find any reference to the term "byte rate" in either of the two
PDF documents that define the original spec. The similar term
"data rate" is used as a descriptor of this field, but not as the
field name.