Enabling site to be built using Ruby 3.2.0
Previously a `NoMethodError` related to `untaint` would be raised
when trying to run the `bundle` command on this Ruby version.
Upgrading to the latest version of the `bundler` gem (2.4.3)
removes this error.

However, the `liquid` gem dependency also has to be upgraded as
well, or a similar error will be raised when running `bundle exec
jekyll serve`.

Since the oldest Ruby version Bundler 2.4.3 supports is Ruby
2.7.0, that is the oldest Ruby version that can be used to build
the site. However, that was already the case before this commit
because of the `rouge` 4.0.1 gem dependency.