More correct explanation of Samples union
Previously it was implied that the first extra field of a
WAVE_FORMAT_EXTENSIBLE "fmt " chunk always represented the number
of valid bits per sample, but that isn't the case. The article
that describes the WAVE_FORMAT_EXTENSIBLE format calls this field
the Samples union, because the C struct used by the Windows API
for WAVE_FORMAT_EXTENSIBLE uses a union for this field.