Detailing WAVE_FORMAT_EXTENSIBLE and extra fields
The previous description implied that it isn't possible to include
the usual extra fields that a non-WAVE_FORMAT_EXTENSIBLE format
might include when using WAVE_FORMAT_EXTENSIBLE, but that isn't
the case. The full WAVE_FORMAT_EXTENSIBLE chunk structure is now
being shown because it seems like the most clear way to show that.