Debug: Reordering methods for better readability
It seems more clear to structure the code such that methods that
implement the main outline of the program are listed first, with
helper methods following. This would imply that the helper methods
that display each line of the output should be placed later in the
program, rather than at the beginning.

In addition to moving these three methods to the end of the
program, `display_chunk_header()` is being placed before
`display_line()`, because it will normally be called earliest of
the two when displaying a chunk in the output.