Debug: Wrapping values longer than column width
Previously, long values in the parsed value or raw bytes column
would extend out of the column, but they will now wrap as needed.
(This isn't an issue for the field name and data type columns
because they never have a value longer than the column width).
This change should cause each line in the output to be a maximum
of 80 characters long, and generally make the output easier to
read.

The parsed value and raw bytes columns have their width adjusted
as part of this change. The parsed value column now shows 18
characters per line for strings and 19 characters for other types,
up from 10 characters per line in un-wrapped output. (The 18 vs.
19 characters discrepency is caused by formatting strings such
that each line after the first is indented, to make it more clear
that the opening " character is not part of the string itself).
The raw bytes column now shows a maximum of 8 bytes (i.e. 16 hex
digits), down from 11 bytes in un-wrapped output. These new column
widths seem to provide a better balance that makes the output
easier to read.