Debug: 32 bytes of "data" chunk shown, up from 10
It seems more generally useful to show a larger amount of the
"data" chunk. 10 bytes was originally chosen so that the content
would not expand out of the column, but now that long values wrap
(as of commit 3db9a639052181d5e2b1100562cdc3f5d175c13f) a larger
number of bytes can be shown.