Debug: Chunk sizes are fully respected
As well as sizes for subsections, such as a "fmt " chunk
extension.

Previously, field reading logic did not take the stated number of
bytes in a chunk or chunk sub section into account, nor the
available space in a parent chunk. This could cause misleading
output. After this commit, chunk and section sizes are strictly
followed.

For example, if a WAVE_FORMAT_EXTENSIBLE "fmt " chunk extension
indicates a size of 4 bytes, now only 4 bytes will be read and
displayed, instead of the normal 22 bytes for this type of chunk
extension. Or, if a "smpl" chunk is the last chunk in a file, and
the file is large enough to contain it, but based on its size it
extends past the end of the parent "RIFF" chunk, only the portion
that fits inside the "RIFF" chunk will be displayed.

Fields that have partial data before being cut off by the end of a
container are shown with the empty bytes displayed as "__". For
example, a 4 byte field where only two bytes could be read might
be displayed as "a3 f1 __ __". However, as before, fields with
partial data due to a premature end of file are still not shown.
To reduce the scope of this commit that is being left as an
improvement for the future.

To implement these changes, `FieldReader` now has the concept of a
"byte limit". A byte limit represents the number of bytes left
until running out of room in a chunk or chunk subsection. A byte
limit can be added by calling `FieldReader#with_byte_limit` with a
block. The byte limit is in place until the block is exited.
Fields are read optimistically inside the block, but if the byte
limit is prematurely reached the block will automatically be
exited and execution will continue in the outer context.

To reflect that data inside a *.wav file is nested, a nested byte
limit can be added by calling `FieldReader#with_byte_limit` within
a parent `with_byte_limit` block. For example, there might
simultaneously be a byte limit for the parent "RIFF" chunk, child
"fmt " chunk, and "fmt " chunk extension. Each byte that is read
counts toward each of the byte limits present. A nested byte limit
is clamped to the size of its parent's remaining budget, to
prevent content from overflowing out of the parent.

The upshot is that by calling `FieldReader#with_byte_limit` in the
appropriate places, the stated sizes of each part of the file can
be strictly followed without field reading code having to manually
keep track of how many bytes have been read, are remaining, etc.