Debug: Renaming `:parsed_value` -> `:value`
The fact that a field's value happens to be parsed from a set of
bytes doesn't seem to be an important distinction to make; there
isn't a corresponding concept of a "non-parsed value". Besides
being more succinct, the term "value" might make the code more
clear by removing possible confusion about what it means exactly
that a value is "parsed".