Debug: Renaming `:type_label` -> `:type`
The `_label` suffix seemed a little confusing since field hashes
also have a plain `label` key. The `_label` suffix was originally
added to indicate that the value represents a string for display
(and hence should not necessarily be used for logic). However, it
_is_ used for logic in `display_field()`, and having a distinction
between an abstract value representing a type and a display string
for that type seems like maybe overkill for the program's current
level of complexity. That might change in the future but in the
meantime the simpler name can be used.