Debug: Showing "__" for missing "extra" bytes
This commit changes how extra bytes at the end of a chunk or chunk
subsection are displayed when the chunk or subsection is truncated
due to the size of its parent container.

For example, suppose the size of a chunk with a defined set of
fields is 4 bytes larger than usual, but its parent chunk's size
only allows for 2 extra bytes. Previously, if these two extra
bytes had values of f0 and b1, they would be displayed as "f0 b1".
Now they will be displayed as "f0 b1 __ __".

This makes the display of extra bytes more consistent with how
other fields that are prematurely truncated are displayed.
Although the extra bytes don't really represent a concrete field
per se, it seems useful to indicate truncation in some way.

Similar to how other fields are displayed, if truncation results
in there being no extra bytes, or if the truncation is due to an
unexpected end of file, then no "extra bytes" field will be
displayed in the output.

To facilitate this change, tracking of byte limits is now
implemented by counting up to the byte limit, rather than down to
0. Due to the greater complexity, a new `ByteLimit` object is used
to represent each byte limit, instead of a raw integer.