Debug: Fields incomplete due to EOF are shown
Previously, if an unexpected end-of-file error prevented some of
the bytes for a field from being read, the field would not be
shown in the output. Now it will be (if at least one byte can be
read), with missing bytes displayed as "__". This matches how
fields that are prematurely cut off due to overflowing the
containing chunk or chunk subsection are displayed, and helps to
give a more accurate view of the file's contents.