More accurate size for "smpl" chunk in fixture
The previous "smpl" chunk size prevented the (truncated) "sampler
specific data" field from being included in the chunk. It was
still present in the file, but after the end of the "RIFF" chunk.

This commit shouldn't change the semantics of the relevant test
because in both the old and new fixture version the "sampler
specific data size" field should be too large. However, it seems
better for the fixture to be set up as actually intended.