Error raised for unsupported fixture sample format
Previously, if a fixture was configured to have non-empty sample
data but the sample format was unsupported, `fixture_writer.rb`
would silently stop writing the fixture *.wav file and no error
would occur. This made it easier for a fixture to be configured
incorrectly without it being known. An explicit error will now
occur in this situation.