Skipping missing values when writing fixtures
Previously, a missing value for most fixture fields would cause
writing data to the fixture *.wav file to stop. However, this was
not a very flexible approach, because it prevented writing a
partial chunk followed by other chunks.

In fact, this caused the `invalid/smpl_chunk_empty.wav` fixture to
be inadvertantly missing the "data" chunk. This caused the fixture
to have more than one source of error, which could have led to an
incorrect test result. (It didn't, but in theory it could have).

After this commit, missing fixture field values will always be
skipped and writing will continue to the next field, except when
attempting to write sample data with an unsupported format.