Removing `channel_format` fixture config field
This field (part of the "data" chunk config) matched the channel
count from the "fmt" chunk config in all fixtures that have sample
data present. Maybe in the future it might be useful to manually
configure the channel count (and sample format, etc.) for cases
where e.g. an invalid fixture has two "fmt " chunks. But until
then it doesn't need to be present, and in those future cases it
can be optional.