Fixing "RIFF" chunk size in fixture
Previously, the "RIFF" chunk size (automatically calculated due to
the "auto" value in the YAML config) was larger than the number of
remaining bytes in the file. This introduced an unintended extra
source of error in the fixture beyond what it was intended to
include. The automatic "RIFF" chunk size calculation only goes off
the sizes that chunks report, so it's not a good fit for fixtures
such as this that have chunks with intentionally wrong sizes.

I don't think the wrong "RIFF" chunk size affected any tests,
because when attempting to open the file the "fmt " chunk
extension being too small causes an error before any EOF error can
occur. It's still good to correct the "RIFF" chunk size for the
future though.

This commit also adds some comments to the YAML config as a future
aid toward understanding how the "RIFF" chunk size is calculated.