Fixing "RIFF" chunk size in fixture
Previously, the "RIFF" chunk size (automatically calculated due to
the "auto" value in the YAML config) was 36, 4 bytes larger than
the actual number of remaining bytes in the file. The reason was
that the automatic "RIFF" chunk size calculation doesn't work
properly for chunks with a partially complete header. Although the
automatic size calculation could be improved, it seems more
straightforward to manually set the "data" chunk size for now.

This shouldn't change the results of any tests because both the
old and new versions of this fixture cause `InvalidFormatError` to
be raised with the same "Unexpected end of file" message when
attempting to read it.