Fixing "RIFF" chunk size in several fixtures
Previously, the "RIFF" chunk size (automatically calculated due to
the "auto" value in the YAML config) in each of these fixtures was
smaller than the number of remaining bytes in the file. This meant
that some bytes intended to be part of the main "RIFF" chunk
instead came after it. The automatic "RIFF" chunk size calculation
only goes off the sizes that chunks report, so it's not a good fit
for fixtures such as these that have chunks with intentionally
wrong sizes.

In practice I don't think this caused any testing issues. When
reading each of these fixtures a problem with the "fmt " chunk
will cause an error to be raised before the end of the file is
reached. However, to prevent multiple sources of error, it seems
like a good idea for the "RIFF" chunk size in each fixture to
accurately reflect the number of bytes present.