Fixing two "RIFF" chunk sizing bugs for fixtures
Previously, if a fixture's YAML config had the "RIFF" chunk size
set to "auto" and there was also a "data" chunk, two bugs in the
automatic sizing calculation cancelled each other out, leading to
an accidentally correct result. This commit fixes the logic to be
correct for the correct reasons, and removes a possible source of
error in the future.

The first bug was that the "RIFF" chunk's FourCC and size fields
were erroneously added to the calculated chunk size, leading it to
being 8 bytes too large. The second bug was that the "data" chunk
FourCC and size fields _weren't_ included, leading to 8 bytes not
being accounted for.

If the fixture had a "data" chunk, the extra 8 bytes and missing 8
bytes cancelled each other out and the "RIFF" chunk size was
correct.

If there wasn't a "data" chunk, the "RIFF" chunk size was 8 bytes
too large. However, that didn't affect any fixtures because there
aren't currently aren't any that have a "RIFF" chunk size of
"auto" and are also missing a "data" chunk.

A known issue is that the calculated "RIFF" chunk size will be
incorrect if the "data" chunk header is partially complete (i.e.
present but less than 8 bytes in size). However, that is a pre-
existing issue and also affects the other chunks types. That
can be fixed in a different commit.