Removing use of global constants for sharing data
Previously, two global constants whose values were determined at
runtime (`SQUARE_WAVE_CYCLE_REPEATS` and `TOTAL_SAMPLE_FRAMES`)
were used to share data between the top-level scope and helper
methods. This commit removes this form of data sharing, in favor
of passing values in method arguments. The two constants are also
being converted to regular variables to limit their visibility.

The main driver for doing this is to allow the top-level code to
be wrapped in a method in a future commit. However, this change
seems like a good idea in general to make the flow of data easier
to understand and the code more maintainable.