Revert "Improving fixture generation time ~8.7x"
This reverts commit 8d9ea990f54f32858648be1557d47e344578a852.

This commit is being reverted because the performance measurements
in the commit message were misleading. Commit
d734e8a5b13f156d20984b021548465fec984d19 is also being reverted
(in a separate revert commit) for the same reason.

When running the performance tests I inadvertantly had an
experimental uncommitted fixture with a large amount of sample
data in my copy of the fixtures folder. This skewed the test
results in both of these commits compared to what would be seen
when only generating the fixtures that are actually in the
repository.

These two revert commits will be followed by the re-application of
the original two commits, with hopefully more accurate performance
measurements in the commit messages this time.