Reading files in tests in Windows-compatible way
I believe that when reading binary files on Windows the "rb" mode
needs to be used instead of "r". `File.read` uses "r" by default,
unlike `File.binread` which uses "rb".

In theory this means that some of the `Reader` tests might have
been broken when using Windows. I say "in theory" because I
currently don't have a Windows machine to test this out on.