Fixing tests to read from the correct files
Previously in both of these two tests the first and second parts
of the test incorrectly read from two different files. Presumably
this was due to copy-and-pasting these tests from existing tests,
and not changing the relevant file name in both required places.

The tests continue to pass after this commit, so presumably there
wasn't a bug accidentally being hidden.