Removing duplicated logic for building file names
Previously the tests that used `exhaustively_test()` were
responsible for constructing the file name of each test fixture.
This led to the logic for this being duplicated in several places.
Also, the `format_chunk_format` block argument was only ever used
for constructing file names, and not directly in any tests.

Now the file name is instead constructed inside of
`exhaustively_test()` and passed to the relevant block, in place
of `format_chunk_format`. This removes duplication and seems to
simplify things.