Using better field value in fixtures
Previously, these fixtures all contained a "smpl" chunk that used
a value of 0 for the `sample_nanoseconds` field. (I.e. the field
called "dwSamplePeriod" in the spec). However, this value doesn't
seem conceptually valid. It seems better to use the value 22675,
which the spec calls out as the normal value to use when the
"fmt " chunk has a sample rate of 44100 (which is the case for all
of these fixtures).

This field value is intentionally not being changed in the
"valid/with_sample_chunk_with_fields_out_of_range.wav" fixture,
because a value of 0 seems to fit with the "invalid but won't
cause an error" theme of that fixture.