Adding tests for files with multiple "smpl" chunks
As far as I'm currently aware the *.wav spec doesn't specifically
say whether or not files with more than one "smpl" chunk are
valid, and if so how they should be handled. The gem currently
will successfully read a file with multiple "smpl" chunks, but
`Reader#sampler_info` will be populated with data from the final
"smpl" chunk, and the earlier instances will be ignored. The new
tests in this commit document this behavior and will hopefully
prevent it from being unintentionally changed in the future.

These fixtures were created using a modified version of
`tools/fixture_writer.rb`, because that tool currently doesn't
have the ability to write a fixture that has more than one chunk
of the same type. Adding that ability in a non-hacky way should be
a relatively large change so it is being deferred until later. As
a result corresponding `*.yml` files aren't being committed.